# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a  Demo Template API with Swagger documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3030",
    "basePath": "/pickup/v1",
    "paths": {
        "/addressdetails/": {
            "post": {
                "description": "create a new address  for a customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Create a new address for a customer",
                "parameters": [
                    {
                        "description": "Create address request",
                        "name": "createAddress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createAddress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "address created",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/addressdetails/{addressid}": {
            "put": {
                "description": "update a address's based on addressid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Update a address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "update address",
                        "name": "addressid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateAddressDetails",
                        "name": "updateAddressDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateAddressDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a address based on addressid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Delete a address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "addressid",
                        "name": "addressid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted succesfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/addressdetails/{customerid}": {
            "get": {
                "description": "get a address by customerid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Get a address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customerid ID",
                        "name": "customerid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/addresstypelist/": {
            "get": {
                "description": "Get a Addresstypelist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get a Addresstypelist",
                "responses": {
                    "200": {
                        "description": "Addresstype list retrieved",
                        "schema": {
                            "$ref": "#/definitions/domain.Addresstype"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/cancel/{pickuprequestid}": {
            "put": {
                "description": "cancel a pickup request based on pickuprequestid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickuprequest"
                ],
                "summary": "cancel a pickup request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cancel pickuprequest",
                        "name": "pickuprequestid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pickuprequest Cancelled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/Assigned/{facilityid}": {
            "get": {
                "description": "Get the Assigned pickuprequest based on facilityid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the Assigned pickuprequest based on facilityid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickupfacilityid",
                        "name": "facilityid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/Unassigned/{facilityid}": {
            "get": {
                "description": "Get the Unassigned pickuprequest based on facilityid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the Unassigned pickuprequest based on facilityid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickupfacilityid",
                        "name": "facilityid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unassigned retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.UnassignedDom"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/count/{facilityid}": {
            "get": {
                "description": "Get the count of unassigned , assigned , pickedup , cancelled , total sum based on facilityid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the count of unassigned , assigned , pickedup , cancelled , total sum based on facilityid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickupfacilityid",
                        "name": "facilityid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "count retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.CountApiResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/customerpickuprequests/{customerid}": {
            "get": {
                "description": "Get the pickuprequest raised by the customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the pickuprequest raised by the custoner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customerid",
                        "name": "customerid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unassigned retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.PickuprequestBasiccus"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/getpickuprequest/{pickupagentid}": {
            "get": {
                "description": "Get the pickuprequest assigned to the pickupagent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the pickuprequest assigned to the pickupagent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pickupagentid",
                        "name": "pickupagentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/pickuprequest/{facilityid}": {
            "get": {
                "responses": {
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Addresstype": {
            "type": "object",
            "properties": {
                "addresstype": {
                    "type": "string"
                },
                "addresstypeid": {
                    "type": "integer"
                }
            }
        },
...
        "handler.createAddress": {
            "type": "object",
            "required": [
                "addressline1",
                "addressline2",
                "addresstype",
                "city",
                "country",
                "customerid",
                "emailid",
                "firstname",
                "fromtopickup",
                "geocode",
                "isverified",
                "landmark",
                "lastname",
                "mobilenumber",
                "pincode",
                "state"
            ],
            "properties": {
                "addressline1": {
                    "type": "string"
                },
                "addressline2": {
                    "type": "string"
                },
                "addresstype": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "customerid": {
                    "type": "string"
                },
                "emailid": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "fromtopickup": {
                    "type": "string"
                },
                "geocode": {
                    "type": "string"
                },
                "isverified": {
                    "type": "boolean"
                },
                "landmark": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "mobilenumber": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    }
}